bot/requirements.txt
kubernetes>=29.0.0
google-generativeai>=0.4.0
requests>=2.31.0